fix(compose): the overlay search door is its own file, and a gate names the surface it checks
Two small corrections, both to things that read as green while being untrue.

overlayread.go reached 503 lines and fails the 500-LOC cap. It crossed in #826
and nothing caught it, because the gate runs over the whole tree and the merge
after it was docs-only — so the change classifier skipped the Go lane and main
went green while latently red. The next Go-touching pull request inherits it.

Split along the seam the file already had rather than by line count. Search is
its own concept: every get/list handler serves ONE entity type through one
provider call, while search walks every mirrored type, holds its own limit
policy, and folds heterogeneous hits into one ranked answer. It is the only read
here that decides what a result MEANS across types, which is why the vocabulary
and the clamp travel with it. 334 lines and 191 rather than 503 and a waiver.

TestEveryDocumentAToolNamesIsOneTheSurfacePublishes named a surface it does not
check. Two surfaces serve these descriptions and only one serves documents: an
MCP client reaches resources/read, and the Surface-B runner has no document seam
at all, so a run reads "published at margince://schema/record-fields" and cannot
fetch it.

The gate is not made to fail over that, because it costs a run nothing: the
runner is offered no resource step form either, so it cannot spend a turn
discovering the pointer is unreachable — the text is inert rather than a dead
end. Giving it the seam was measured and made things worse (#737), so the
asymmetry is a decision. What would make it an oversight is a reader believing
this gate covers both surfaces, so the name says HOSTED and the comment says why.
